Add CustomRowsSyncTask scheduled task and shared CustomRowFetchService
Scan configured custom home sections across user profiles and proactively
refresh Letterboxd, TMDB, and MDBList custom user lists daily at 4:00 AM.
