+++ Jellyfin/backend/Api/CustomRowController.cs
@@ -1,10 +1,10 @@
-using System.Collections.Concurrent;
-using System.IO;
-using System.Net.Http;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using System.Text.RegularExpressions;
-using System.Xml.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -17,34 +17,20 @@ namespace Moonfin.Server.Api;
 [Route("Moonfin/CustomRows")]
 public class CustomRowController : ControllerBase
 {
-    private readonly MoonfinSettingsService _settingsService;
     private readonly CustomRowCacheService _cacheService;
-    private readonly ImdbListsCacheService _imdbCacheService;
-    private readonly IHttpClientFactory _httpClientFactory;
+    private readonly CustomRowFetchService _fetchService;
     private readonly ILogger<CustomRowController> _logger;
-    private readonly ILogger<ImdbListsTask> _taskLogger;
 
     private static readonly TimeSpan CacheTtl = TimeSpan.FromHours(24);
 
-    // Page through MDBList user lists 100 at a time and stop at 500 so a huge
-    // list can't burn the whole API quota.
-    private const int MdbListPageSize = 100;
-    private const int MdbListMaxItems = 500;
-
     public CustomRowController(
-        MoonfinSettingsService settingsService,
         CustomRowCacheService cacheService,
-        ImdbListsCacheService imdbCacheService,
-        IHttpClientFactory httpClientFactory,
-        ILogger<CustomRowController> logger,
-        ILogger<ImdbListsTask> taskLogger)
+        CustomRowFetchService fetchService,
+        ILogger<CustomRowController> logger)
     {
-        _settingsService = settingsService;
         _cacheService = cacheService;
-        _imdbCacheService = imdbCacheService;
-        _httpClientFactory = httpClientFactory;
+        _fetchService = fetchService;
         _logger = logger;
-        _taskLogger = taskLogger;
     }
 
     [HttpGet("Items")]
@@ -72,24 +58,6 @@ public class CustomRowController : ControllerBase
             return Unauthorized(new { Error = "User not authenticated" });
         }
 
-        if (source == "imdb")
-        {
-            try
-            {
-                var imdbItems = await FetchImdbList(type, cancellationToken);
-                return Ok(new CustomRowResponse
-                {
-                    Success = true,
-                    Items = imdbItems
-                });
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Failed to resolve IMDb custom row for type: {Type}", type);
-                return StatusCode(StatusCodes.Status500InternalServerError, new { Error = ex.Message });
-            }
-        }
-
         Dictionary<string, string> parsedParams;
         try
         {
@@ -110,10 +78,7 @@ public class CustomRowController : ControllerBase
                 .ToDictionary(kv => kv.Key, kv => kv.Value);
         }
 
-        var canonicalParams = JsonSerializer.Serialize(
-            new SortedDictionary<string, string>(parsedParams, StringComparer.Ordinal));
-        var paramHash = GetStringSha256Hash(canonicalParams);
-        var cacheKey = $"{source}:{type}:{paramHash}";
+        var cacheKey = CustomRowFetchService.ComputeCacheKey(source, type, parsedParams);
 
         if (!refresh)
         {
@@ -130,25 +95,7 @@ public class CustomRowController : ControllerBase
 
         try
         {
-            List<CustomRowItem> items = new();
-
-            switch (source)
-            {
-                case "mdblist":
-                    items = await FetchMdbList(type, parsedParams, userId.Value, cancellationToken);
-                    break;
-                case "tmdb":
-                    items = await FetchTmdb(type, parsedParams, userId.Value, cancellationToken);
-                    break;
-                case "tmdb_chart":
-                    items = await FetchTmdbChart(type, userId.Value, cancellationToken);
-                    break;
-                case "letterboxd":
-                    items = await FetchLetterboxd(type, parsedParams, userId.Value, cancellationToken);
-                    break;
-                default:
-                    return BadRequest(new { Error = $"Unsupported custom row source: {source}" });
-            }
+            var items = await _fetchService.FetchCustomRowAsync(source, type, parsedParams, userId.Value, cancellationToken);
 
             // Don't cache empty results. An empty row cached for 24h looks like a
             // broken list when the real cause was an upstream hiccup.
@@ -175,694 +122,6 @@ public class CustomRowController : ControllerBase
             });
         }
     }
-
-    private async Task<List<CustomRowItem>> FetchMdbList(
-        string type,
-        Dictionary<string, string> paramsMap,
-        Guid userId,
-        CancellationToken cancellationToken)
-    {
-        var resolved = await _settingsService.GetResolvedProfileAsync(userId, "global");
-        var apiKey = resolved?.MdblistApiKey;
-        if (string.IsNullOrWhiteSpace(apiKey))
-        {
-            throw new InvalidOperationException("MDBList API key is not configured.");
-        }
-
-        paramsMap.TryGetValue("username", out var username);
-        paramsMap.TryGetValue("listname", out var listname);
-
-        if (string.IsNullOrWhiteSpace(username) || string.IsNullOrWhiteSpace(listname))
-        {
-            throw new ArgumentException("MDBList requires both username and listname parameters.");
-        }
-
-        var baseUrl = $"https://api.mdblist.com/lists/{Uri.EscapeDataString(username)}/{Uri.EscapeDataString(listname)}/items?apikey={Uri.EscapeDataString(apiKey)}&limit={MdbListPageSize}&append_to_response=poster";
-        var client = CreateClient();
-
-        var items = new List<CustomRowItem>();
-        string? cursor = null;
-
-        // Follow pagination.next_cursor until it's absent, that's the only
-        // continuation signal the MDBList API guarantees.
-        while (items.Count < MdbListMaxItems)
-        {
-            var url = cursor == null ? baseUrl : $"{baseUrl}&cursor={Uri.EscapeDataString(cursor)}";
-
-            using var response = await client.GetAsync(url, cancellationToken).ConfigureAwait(false);
-            if ((int)response.StatusCode == 429)
-            {
-                throw new Exception("MDBList rate limit reached. Try again later.");
-            }
-
-            if (!response.IsSuccessStatusCode)
-            {
-                throw new Exception($"MDBList API returned status {(int)response.StatusCode}");
-            }
-
-            var json = await response.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-            using var document = JsonDocument.Parse(json);
-            var root = document.RootElement;
-
-            var countBefore = items.Count;
-
-            if (root.ValueKind == JsonValueKind.Array)
-            {
-                // A flat array is a single page with no pagination object.
-                AppendMdbListItems(root, items);
-                break;
-            }
-
-            if (root.ValueKind != JsonValueKind.Object)
-            {
-                break;
-            }
-
-            if (root.TryGetProperty("movies", out var moviesProp) && moviesProp.ValueKind == JsonValueKind.Array)
-            {
-                AppendMdbListItems(moviesProp, items);
-            }
-
-            if (root.TryGetProperty("shows", out var showsProp) && showsProp.ValueKind == JsonValueKind.Array)
-            {
-                AppendMdbListItems(showsProp, items);
-            }
-
-            cursor = null;
-            if (root.TryGetProperty("pagination", out var paginationProp)
-                && paginationProp.ValueKind == JsonValueKind.Object
-                && paginationProp.TryGetProperty("next_cursor", out var cursorProp)
-                && cursorProp.ValueKind == JsonValueKind.String)
-            {
-                cursor = cursorProp.GetString();
-            }
-
-            if (string.IsNullOrEmpty(cursor) || items.Count == countBefore)
-            {
-                break;
-            }
-        }
-
-        // Movies and shows arrive as separate arrays per page, so restore true list order.
-        items = items
-            .OrderBy(i => i.Rank ?? int.MaxValue)
-            .Take(MdbListMaxItems)
-            .ToList();
-        for (var i = 0; i < items.Count; i++)
-        {
-            items[i].Rank = i + 1;
-        }
-
-        // Fetch backdrops (and any posters MDBList didn't provide) from TMDb in parallel
-        var resolvedProfile = await _settingsService.GetResolvedProfileAsync(userId, "global");
-        var tmdbKey = resolvedProfile?.TmdbApiKey;
-        if (!string.IsNullOrWhiteSpace(tmdbKey))
-        {
-            using var tmdbSemaphore = new SemaphoreSlim(15, 15);
-            var movieTasks = items.Where(i => i.Id.HasValue).Select(async rowItem =>
-            {
-                await tmdbSemaphore.WaitAsync(cancellationToken).ConfigureAwait(false);
-                try
-                {
-                    var isShow = rowItem.Type == "Series";
-                    var tmdbUrl = isShow 
-                        ? $"https://api.themoviedb.org/3/tv/{rowItem.Id}"
-                        : $"https://api.themoviedb.org/3/movie/{rowItem.Id}";
-
-                    using var req = new HttpRequestMessage(HttpMethod.Get, tmdbUrl);
-                    TmdbRequestHelper.ApplyAuth(req, tmdbKey);
-                    using var resp = await client.SendAsync(req, cancellationToken).ConfigureAwait(false);
-                    if (resp.IsSuccessStatusCode)
-                    {
-                        var detailsJson = await resp.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-                        using var doc = JsonDocument.Parse(detailsJson);
-                        var detailsRoot = doc.RootElement;
-                        
-                        if (string.IsNullOrEmpty(rowItem.PosterUrl)
-                            && detailsRoot.TryGetProperty("poster_path", out var pProp) && pProp.ValueKind == JsonValueKind.String)
-                        {
-                            rowItem.PosterUrl = pProp.GetString();
-                        }
-                        if (detailsRoot.TryGetProperty("backdrop_path", out var bProp) && bProp.ValueKind == JsonValueKind.String)
-                        {
-                            rowItem.BackdropUrl = bProp.GetString();
-                        }
-                    }
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogWarning(ex, "Failed to fetch TMDb details for MDBList item ID: {Id}", rowItem.Id);
-                }
-                finally
-                {
-                    tmdbSemaphore.Release();
-                }
-            });
-            await Task.WhenAll(movieTasks).ConfigureAwait(false);
-        }
-
-        return items;
-    }
-
-    /// <summary>
-    /// Parses one MDBList list-items array (movies or shows) into <paramref name="items"/>.
-    /// Keeps each item's own MDBList rank so movies and shows can be re-merged in list order.
-    /// </summary>
-    private static void AppendMdbListItems(JsonElement itemsArray, List<CustomRowItem> items)
-    {
-        foreach (var item in itemsArray.EnumerateArray())
-        {
-            string? imdbId = null;
-            string? tmdbId = null;
-
-            if (item.TryGetProperty("ids", out var idsObj) && idsObj.ValueKind == JsonValueKind.Object)
-            {
-                if (idsObj.TryGetProperty("imdb", out var imdbProp) && imdbProp.ValueKind == JsonValueKind.String)
-                {
-                    imdbId = imdbProp.GetString();
-                }
-
-                if (idsObj.TryGetProperty("tmdb", out var tmdbProp))
-                {
-                    tmdbId = tmdbProp.ValueKind == JsonValueKind.Number
-                        ? tmdbProp.GetInt64().ToString()
-                        : tmdbProp.ValueKind == JsonValueKind.String ? tmdbProp.GetString() : null;
-                }
-            }
-
-            if (string.IsNullOrWhiteSpace(imdbId) && item.TryGetProperty("imdb_id", out var imdbIdProp) && imdbIdProp.ValueKind == JsonValueKind.String)
-            {
-                imdbId = imdbIdProp.GetString();
-            }
-
-            if (string.IsNullOrWhiteSpace(tmdbId) && item.TryGetProperty("id", out var idProp) && idProp.ValueKind == JsonValueKind.Number)
-            {
-                tmdbId = idProp.GetInt64().ToString();
-            }
-
-            var title = item.TryGetProperty("title", out var titleProp) ? titleProp.GetString() ?? "Unknown" : "Unknown";
-
-            int? year = null;
-            if (item.TryGetProperty("release_year", out var yrProp) && yrProp.ValueKind == JsonValueKind.Number)
-            {
-                year = yrProp.GetInt32();
-            }
-
-            var mediaType = item.TryGetProperty("mediatype", out var mediaProp) ? mediaProp.GetString()?.ToLowerInvariant() : null;
-            var finalType = (mediaType == "show" || mediaType == "shows" || mediaType == "series" || mediaType == "tv") ? "Series" : "Movie";
-
-            int? rank = null;
-            if (item.TryGetProperty("rank", out var rankProp) && rankProp.ValueKind == JsonValueKind.Number)
-            {
-                rank = rankProp.GetInt32();
-            }
-
-            string? posterUrl = null;
-            if (item.TryGetProperty("poster", out var pProp) && pProp.ValueKind == JsonValueKind.String)
-            {
-                posterUrl = MdbListApiHelper.NormalizeTmdbImagePath(pProp.GetString());
-            }
-
-            items.Add(new CustomRowItem
-            {
-                Id = long.TryParse(tmdbId, out var parsedTmdbId) ? parsedTmdbId : null,
-                Name = title,
-                Type = finalType,
-                ProductionYear = year,
-                Rank = rank,
-                ProviderIds = new CustomRowItemProviderIds
-                {
-                    Imdb = imdbId,
-                    Tmdb = tmdbId
-                },
-                PosterUrl = posterUrl
-            });
-        }
-    }
-
-    private async Task<List<CustomRowItem>> FetchTmdb(
-        string type,
-        Dictionary<string, string> paramsMap,
-        Guid userId,
-        CancellationToken cancellationToken)
-    {
-        var resolved = await _settingsService.GetResolvedProfileAsync(userId, "global");
-        var apiKey = resolved?.TmdbApiKey;
-        if (string.IsNullOrWhiteSpace(apiKey))
-        {
-            throw new InvalidOperationException("TMDB API Key is not configured.");
-        }
-
-        paramsMap.TryGetValue("id", out var listId);
-        if (string.IsNullOrWhiteSpace(listId))
-        {
-            throw new ArgumentException("TMDB requires id parameter.");
-        }
-
-        var isCollection = type == "movie_collection";
-        var url = isCollection
-            ? $"https://api.themoviedb.org/3/collection/{Uri.EscapeDataString(listId)}"
-            : $"https://api.themoviedb.org/3/list/{Uri.EscapeDataString(listId)}";
-
-        var client = CreateClient();
-        using var request = new HttpRequestMessage(HttpMethod.Get, url);
-        TmdbRequestHelper.ApplyAuth(request, apiKey);
-
-        using var response = await client.SendAsync(request, cancellationToken).ConfigureAwait(false);
-        if (!response.IsSuccessStatusCode)
-        {
-            throw new Exception($"TMDB API returned status {(int)response.StatusCode}");
-        }
-
-        var json = await response.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-        using var document = JsonDocument.Parse(json);
-        var root = document.RootElement;
-
-        var items = new List<CustomRowItem>();
-        int rank = 1;
-
-        if (isCollection)
-        {
-            if (root.TryGetProperty("parts", out var partsProp) && partsProp.ValueKind == JsonValueKind.Array)
-            {
-                foreach (var part in partsProp.EnumerateArray())
-                {
-                    var partId = part.TryGetProperty("id", out var idProp) ? idProp.GetInt64().ToString() : null;
-                    var title = part.TryGetProperty("title", out var titleProp) ? titleProp.GetString() ?? "Unknown" : "Unknown";
-                    var dateStr = part.TryGetProperty("release_date", out var rdProp) ? rdProp.GetString() : null;
-                    int? year = null;
-                    if (!string.IsNullOrEmpty(dateStr) && dateStr.Length >= 4 && int.TryParse(dateStr.Substring(0, 4), out var yr))
-                    {
-                        year = yr;
-                    }
-
-                    if (!string.IsNullOrEmpty(partId))
-                    {
-                        var posterPath = part.TryGetProperty("poster_path", out var pProp) ? pProp.GetString() : null;
-                        var backdropPath = part.TryGetProperty("backdrop_path", out var bProp) ? bProp.GetString() : null;
-                        items.Add(new CustomRowItem
-                        {
-                            Id = long.TryParse(partId, out var lbTmdbId) ? lbTmdbId : null,
-                            Name = title,
-                            Type = "Movie",
-                            ProductionYear = year,
-                            Rank = rank++,
-                            ProviderIds = new CustomRowItemProviderIds
-                            {
-                                Tmdb = partId
-                            },
-                            PosterUrl = posterPath,
-                            BackdropUrl = backdropPath
-                        });
-                    }
-                }
-            }
-        }
-        else
-        {
-            if (root.TryGetProperty("items", out var itemsProp) && itemsProp.ValueKind == JsonValueKind.Array)
-            {
-                foreach (var item in itemsProp.EnumerateArray())
-                {
-                    var itemId = item.TryGetProperty("id", out var idProp) ? idProp.GetInt64().ToString() : null;
-                    var title = item.TryGetProperty("title", out var titleProp) ? titleProp.GetString() : null;
-                    if (string.IsNullOrEmpty(title) && item.TryGetProperty("name", out var nameProp))
-                    {
-                        title = nameProp.GetString();
-                    }
-                    title ??= "Unknown";
-
-                    var dateStr = item.TryGetProperty("release_date", out var rdProp) ? rdProp.GetString() : null;
-                    if (string.IsNullOrEmpty(dateStr) && item.TryGetProperty("first_air_date", out var fadProp))
-                    {
-                        dateStr = fadProp.GetString();
-                    }
-
-                    int? year = null;
-                    if (!string.IsNullOrEmpty(dateStr) && dateStr.Length >= 4 && int.TryParse(dateStr.Substring(0, 4), out var yr))
-                    {
-                        year = yr;
-                    }
-
-                    var mediaType = item.TryGetProperty("media_type", out var mtProp) ? mtProp.GetString() : null;
-                    var finalType = mediaType == "tv" ? "Series" : "Movie";
-
-                    if (!string.IsNullOrEmpty(itemId))
-                    {
-                        var posterPath = item.TryGetProperty("poster_path", out var pProp) ? pProp.GetString() : null;
-                        var backdropPath = item.TryGetProperty("backdrop_path", out var bProp) ? bProp.GetString() : null;
-                        items.Add(new CustomRowItem
-                        {
-                            Id = long.TryParse(itemId, out var lbTmdbId) ? lbTmdbId : null,
-                            Name = title,
-                            Type = finalType,
-                            ProductionYear = year,
-                            Rank = rank++,
-                            ProviderIds = new CustomRowItemProviderIds
-                            {
-                                Tmdb = itemId
-                            },
-                            PosterUrl = posterPath,
-                            BackdropUrl = backdropPath
-                        });
-                    }
-                }
-            }
-        }
-
-        return items;
-    }
-
-    private async Task<List<CustomRowItem>> FetchTmdbChart(
-        string type,
-        Guid userId,
-        CancellationToken cancellationToken)
-    {
-        var resolved = await _settingsService.GetResolvedProfileAsync(userId, "global");
-        var apiKey = resolved?.TmdbApiKey;
-        if (string.IsNullOrWhiteSpace(apiKey))
-        {
-            throw new InvalidOperationException("TMDB API Key is not configured.");
-        }
-
-        // type is the API path, e.g. movie/popular, trending/movie/day, etc.
-        var url = $"https://api.themoviedb.org/3/{type}";
-        var client = CreateClient();
-        using var request = new HttpRequestMessage(HttpMethod.Get, url);
-        TmdbRequestHelper.ApplyAuth(request, apiKey);
-
-        using var response = await client.SendAsync(request, cancellationToken).ConfigureAwait(false);
-        if (!response.IsSuccessStatusCode)
-        {
-            throw new Exception($"TMDB API returned status {(int)response.StatusCode} for chart {type}");
-        }
-
-        var json = await response.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-        using var document = JsonDocument.Parse(json);
-        var root = document.RootElement;
-
-        var items = new List<CustomRowItem>();
-        int rank = 1;
-
-        if (root.TryGetProperty("results", out var resultsProp) && resultsProp.ValueKind == JsonValueKind.Array)
-        {
-            foreach (var item in resultsProp.EnumerateArray())
-            {
-                var itemId = item.TryGetProperty("id", out var idProp) ? idProp.GetInt64().ToString() : null;
-                var title = item.TryGetProperty("title", out var titleProp) ? titleProp.GetString() : null;
-                if (string.IsNullOrEmpty(title) && item.TryGetProperty("name", out var nameProp))
-                {
-                    title = nameProp.GetString();
-                }
-                title ??= "Unknown";
-
-                var dateStr = item.TryGetProperty("release_date", out var rdProp) ? rdProp.GetString() : null;
-                if (string.IsNullOrEmpty(dateStr) && item.TryGetProperty("first_air_date", out var fadProp))
-                {
-                    dateStr = fadProp.GetString();
-                }
-
-                int? year = null;
-                if (!string.IsNullOrEmpty(dateStr) && dateStr.Length >= 4 && int.TryParse(dateStr.Substring(0, 4), out var yr))
-                {
-                    year = yr;
-                }
-
-                var mediaType = item.TryGetProperty("media_type", out var mtProp) ? mtProp.GetString() : null;
-                if (string.IsNullOrEmpty(mediaType))
-                {
-                    mediaType = type.Contains("tv") ? "tv" : "movie";
-                }
-                var finalType = mediaType == "tv" ? "Series" : "Movie";
-
-                if (!string.IsNullOrEmpty(itemId))
-                {
-                    var posterPath = item.TryGetProperty("poster_path", out var pProp) ? pProp.GetString() : null;
-                    var backdropPath = item.TryGetProperty("backdrop_path", out var bProp) ? bProp.GetString() : null;
-                    items.Add(new CustomRowItem
-                    {
-                        Id = long.TryParse(itemId, out var lbTmdbId) ? lbTmdbId : null,
-                        Name = title,
-                        Type = finalType,
-                        ProductionYear = year,
-                        Rank = rank++,
-                        ProviderIds = new CustomRowItemProviderIds
-                        {
-                            Tmdb = itemId
-                        },
-                        PosterUrl = posterPath,
-                        BackdropUrl = backdropPath
-                    });
-                }
-            }
-        }
-
-        return items;
-    }
-
-    private async Task<List<CustomRowItem>> FetchLetterboxd(
-        string type,
-        Dictionary<string, string> paramsMap,
-        Guid userId,
-        CancellationToken cancellationToken)
-    {
-        paramsMap.TryGetValue("user", out var username);
-        paramsMap.TryGetValue("name", out var listname);
-        paramsMap.TryGetValue("url", out var fullUrl);
-
-        if (type == "user_list" || type == "watchlist" || type == "films")
-        {
-            throw new ArgumentException("Direct HTML scraping of Letterboxd watchlists and lists is disabled due to their Terms of Service. Please import your list into MDBList and use the MDBList custom row source instead.");
-        }
-
-        if (username != null)
-        {
-            username = username.ToLowerInvariant().Trim();
-        }
-
-        if (string.IsNullOrWhiteSpace(username))
-        {
-            throw new ArgumentException("Letterboxd requires user parameter.");
-        }
-
-        var baseUrl = $"https://letterboxd.com/{Uri.EscapeDataString(username)}/rss/";
-
-        var client = CreateClient();
-        var items = new List<CustomRowItem>();
-
-        var parsedFeedItems = new List<LetterboxdFeedItem>();
-
-        using var response = await client.GetAsync(baseUrl, cancellationToken).ConfigureAwait(false);
-        if (!response.IsSuccessStatusCode)
-        {
-            throw new Exception($"Letterboxd returned status {(int)response.StatusCode} for {baseUrl}");
-        }
-        var xmlContent = await response.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-        var document = XDocument.Parse(xmlContent);
-        var rssItems = document.Descendants("item");
-
-        foreach (var rssItem in rssItems)
-        {
-            var title = rssItem.Element("title")?.Value ?? "Unknown";
-            var link = rssItem.Element("link")?.Value ?? "";
-            
-            var filmTitle = rssItem.Elements().FirstOrDefault(e => e.Name.LocalName == "filmTitle")?.Value ?? title;
-            var filmYearStr = rssItem.Elements().FirstOrDefault(e => e.Name.LocalName == "filmYear")?.Value;
-            var memberRatingStr = rssItem.Elements().FirstOrDefault(e => e.Name.LocalName == "memberRating")?.Value;
-            var resolvedTmdbId = rssItem.Elements().FirstOrDefault(e => e.Name.LocalName == "movieId")?.Value;
-
-            int? year = null;
-            if (int.TryParse(filmYearStr, out var y)) year = y;
-
-            double? rating = null;
-            if (double.TryParse(memberRatingStr, out var r)) rating = r;
-
-            var slugMatch = Regex.Match(link, @"film/([^/]+)/?");
-            if (slugMatch.Success)
-            {
-                var slug = slugMatch.Groups[1].Value;
-                parsedFeedItems.Add(new LetterboxdFeedItem
-                {
-                    Title = filmTitle,
-                    Year = year,
-                    Rating = rating,
-                    Slug = slug,
-                    TmdbId = resolvedTmdbId
-                });
-            }
-        }
-
-        // Fetch full details/posters from TMDb API in parallel (max 15 concurrent requests)
-        var resolvedProfile = await _settingsService.GetResolvedProfileAsync(userId, "global");
-        var tmdbKey = resolvedProfile?.TmdbApiKey;
-        if (!string.IsNullOrWhiteSpace(tmdbKey))
-        {
-            using var tmdbSemaphore = new SemaphoreSlim(15, 15);
-            var movieTasks = parsedFeedItems.Where(i => !string.IsNullOrEmpty(i.TmdbId)).Select(async pItem =>
-            {
-                await tmdbSemaphore.WaitAsync(cancellationToken).ConfigureAwait(false);
-                try
-                {
-                    var tmdbUrl = $"https://api.themoviedb.org/3/movie/{pItem.TmdbId}";
-                    using var req = new HttpRequestMessage(HttpMethod.Get, tmdbUrl);
-                    TmdbRequestHelper.ApplyAuth(req, tmdbKey);
-                    using var resp = await client.SendAsync(req, cancellationToken).ConfigureAwait(false);
-                    if (resp.IsSuccessStatusCode)
-                    {
-                        var detailsJson = await resp.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-                        using var doc = JsonDocument.Parse(detailsJson);
-                        var detailsRoot = doc.RootElement;
-                        if (detailsRoot.TryGetProperty("title", out var titleProp) && titleProp.ValueKind == JsonValueKind.String)
-                        {
-                            pItem.Title = titleProp.GetString() ?? pItem.Title;
-                        }
-                        if (detailsRoot.TryGetProperty("release_date", out var rdProp) && rdProp.ValueKind == JsonValueKind.String)
-                        {
-                            var dateStr = rdProp.GetString();
-                            if (!string.IsNullOrEmpty(dateStr) && dateStr.Length >= 4 && int.TryParse(dateStr.Substring(0, 4), out var yr))
-                            {
-                                pItem.Year = yr;
-                            }
-                        }
-                        if (detailsRoot.TryGetProperty("poster_path", out var pProp) && pProp.ValueKind == JsonValueKind.String)
-                        {
-                            pItem.PosterUrl = pProp.GetString();
-                        }
-                        if (detailsRoot.TryGetProperty("backdrop_path", out var bProp) && bProp.ValueKind == JsonValueKind.String)
-                        {
-                            pItem.BackdropUrl = bProp.GetString();
-                        }
-                    }
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogWarning(ex, "Failed to fetch TMDb details for movie ID: {Id}", pItem.TmdbId);
-                }
-                finally
-                {
-                    tmdbSemaphore.Release();
-                }
-            });
-            await Task.WhenAll(movieTasks).ConfigureAwait(false);
-        }
-
-        // 3. For successfully resolved TMDB IDs, build CustomRowItem lists
-        int rank = 1;
-        foreach (var pItem in parsedFeedItems)
-        {
-            if (!string.IsNullOrEmpty(pItem.TmdbId))
-            {
-                var stars = pItem.Rating.HasValue ? FormatRatingToStars(pItem.Rating.Value) : null;
-                items.Add(new CustomRowItem
-                {
-                    Id = long.TryParse(pItem.TmdbId, out var lbTmdbId) ? lbTmdbId : null,
-                    Name = pItem.Title,
-                    Type = "Movie", // Letterboxd is strictly movies
-                    ProductionYear = pItem.Year,
-                    Rank = rank++,
-                    ProviderIds = new CustomRowItemProviderIds
-                    {
-                        Tmdb = pItem.TmdbId
-                    },
-                    UserRating = stars,
-                    Rating = pItem.Rating,
-                    PosterUrl = pItem.PosterUrl,
-                    BackdropUrl = pItem.BackdropUrl
-                });
-            }
-        }
-
-        return items;
-    }
-
-    private static string FormatRatingToStars(double rating)
-    {
-        int fullStars = (int)Math.Floor(rating);
-        bool halfStar = (rating - fullStars) >= 0.25;
-        var stars = new string('★', fullStars);
-        if (halfStar) stars += "½";
-        return stars;
-    }
-
-
-
-    private HttpClient CreateClient()
-    {
-        var client = _httpClientFactory.CreateClient();
-        client.Timeout = TimeSpan.FromSeconds(15);
-        client.DefaultRequestHeaders.UserAgent.ParseAdd("Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36");
-        return client;
-    }
-
-    // One gate per chart so a burst of clients on an expired chart triggers a single scrape
-    // instead of one per request. Only the handful of chart types ever get added, so the map
-    // doesn't grow unbounded.
-    private static readonly ConcurrentDictionary<string, SemaphoreSlim> ImdbFetchGates =
-        new(StringComparer.OrdinalIgnoreCase);
-
-    private async Task<List<CustomRowItem>> FetchImdbList(string type, CancellationToken cancellationToken)
-    {
-        var cached = _imdbCacheService.TryGetItems(type, TimeSpan.FromDays(1));
-        if (cached != null && cached.Count > 0)
-        {
-            return cached;
-        }
-
-        var gate = ImdbFetchGates.GetOrAdd(type, _ => new SemaphoreSlim(1, 1));
-        await gate.WaitAsync(cancellationToken).ConfigureAwait(false);
-        try
-        {
-            // Another request may have refilled the cache while we waited on the gate.
-            var refreshed = _imdbCacheService.TryGetItems(type, TimeSpan.FromDays(1));
-            if (refreshed != null && refreshed.Count > 0)
-            {
-                return refreshed;
-            }
-
-            _logger.LogInformation("IMDb chart {Type} cache miss or expired, fetching on-demand", type);
-            try
-            {
-                var task = new ImdbListsTask(_httpClientFactory, _imdbCacheService, _taskLogger);
-                var items = await task.FetchChartAsync(type, cancellationToken);
-                if (items != null && items.Count > 0)
-                {
-                    _imdbCacheService.SetItems(type, items);
-                    await _imdbCacheService.FlushAsync();
-                    return items;
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Failed to fetch IMDb chart {Type} on-demand", type);
-            }
-
-            return _imdbCacheService.TryGetItems(type, TimeSpan.FromDays(30)) ?? new List<CustomRowItem>();
-        }
-        finally
-        {
-            gate.Release();
-        }
-    }
-
-    private static string GetStringSha256Hash(string text)
-    {
-        if (string.IsNullOrEmpty(text)) return string.Empty;
-        using var sha = System.Security.Cryptography.SHA256.Create();
-        var textData = System.Text.Encoding.UTF8.GetBytes(text);
-        var hashData = sha.ComputeHash(textData);
-        return Convert.ToHexString(hashData);
-    }
-}
-
-internal class LetterboxdFeedItem
-{
-    public string Title { get; set; } = string.Empty;
-    public int? Year { get; set; }
-    public double? Rating { get; set; }
-    public string Slug { get; set; } = string.Empty;
-    public string? TmdbId { get; set; }
-    public string? PosterUrl { get; set; }
-    public string? BackdropUrl { get; set; }
 }
 
 public class CustomRowResponse
